# docs/api.yml
openapi: 3.0.0
info:
  title: 'WeDigital Garden - Core API'
  description: 'Documentation of Core API for We Digital Garden'
  version: 0.0.1
servers:
  -
    url: 'http://localhost'
paths:
...
  '/api/expertises/{lang}':
    get:
      tags:
        - Expertise
      summary: 'Display a collection of expertises with specific language ressources.'
      operationId: 'App\Http\Controllers\Api\ExpertiseController::getAllWithLang'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: lang
          in: path
          description: 'String for select a specific language'
          required: true
      responses:
        '200':
          description: 'A expertises collection'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/expertise_with_lang_collection_resource'
        '404':
          description: 'A expertises collection not found'
...
  '/api/expertises/{id}/{lang}':
    get:
      tags:
        - Expertise
      summary: 'Display the specified expertise with specific language resource.'
      operationId: 'App\Http\Controllers\Api\ExpertiseController::getOneWithLang'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: id
          in: path
          description: 'Numeric ID of the office to get'
          required: true
          schema:
            $ref: '#/components/schemas/resource_id'
        -
          name: lang
          in: path
          description: 'String for select a specific language'
          required: true
      responses:
        '200':
          description: 'One expertise'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/expertise_with_lang_resource'
        '404':
          description: 'A expertise not found'
...
  /api/formules:
    get:
      tags:
        - Formule
      summary: 'Display a collection of formules ressources.'
      operationId: 'App\Http\Controllers\Api\FormuleController::index'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
      responses:
        '200':
          description: 'A formules collection'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/formule_collection_resource'
        '404':
          description: 'A formules collection not found'
  '/api/formules/{lang}':
    get:
      tags:
        - Formule
      summary: 'Display a collection of formules with specific language ressources.'
      operationId: 'App\Http\Controllers\Api\FormuleController::getAllWithLang'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: lang
          in: path
          description: 'String for select a specific language'
          required: true
      responses:
        '200':
          description: 'A formules collection'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/formule_with_lang_collection_resource'
        '404':
          description: 'A formules collection not found'
  '/api/formules/{id}':
    get:
      tags:
        - Formule
      summary: 'Display the specified formule resource.'
      operationId: 'App\Http\Controllers\Api\FormuleController::getOne'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: id
          in: path
          description: 'Numeric ID of the office to get'
          required: true
          schema:
            $ref: '#/components/schemas/resource_id'
      responses:
        '200':
          description: 'One formule'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/formule_resource'
        '404':
          description: 'A formule not found'
  '/api/formules/{id}/{lang}':
    get:
      tags:
        - Formule
      summary: 'Display the specified formule with specific language resource.'
      operationId: 'App\Http\Controllers\Api\FormuleController::getOneWithLang'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: id
          in: path
          description: 'Numeric ID of the office to get'
          required: true
          schema:
            $ref: '#/components/schemas/resource_id'
        -
          name: lang
          in: path
          description: 'String for select a specific language'
          required: true
      responses:
        '200':
          description: 'One formule'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/formule_with_lang_resource'
        '404':
          description: 'A formule not found'
...
  '/api/framework/certifications/{lang}':
    get:
      tags:
        - Framework
      summary: 'Display a collection of Framework certifications with specific language ressources.'
      operationId: 'App\Http\Controllers\Api\Framework\CertificationController::getAllWithLang'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: lang
          in: path
          description: 'String for select a specific language'
          required: true
      responses:
        '200':
          description: 'A Framework certifications collection'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/certification_with_lang_collection_resource'
        '404':
          description: 'A Framework certifications collection not found'
...
  '/api/framework/certifications/{id}/{lang}':
    get:
      tags:
        - Framework
      summary: 'Display the specified framework certification with specific language resource.'
      operationId: 'App\Http\Controllers\Api\Framework\CertificationController::getOneWithLang'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: id
          in: path
          description: 'Numeric ID of the framework certification to get'
          required: true
          schema:
            $ref: '#/components/schemas/resource_id'
        -
          name: lang
          in: path
          description: 'String for select a specific language'
          required: true
      responses:
        '200':
          description: 'One framework certification'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/certification_with_lang_resource'
        '404':
          description: 'A framework certification not found'
...
  '/api/framework/values/{lang}':
    get:
      tags:
        - Framework
      summary: 'Display a collection of Framework values with specific language ressources.'
      operationId: 'App\Http\Controllers\Api\Framework\ValueController::getAllWithLang'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: lang
          in: path
          description: 'String for select a specific language'
          required: true
      responses:
        '200':
          description: 'A Framework values collection'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/value_with_lang_collection_resource'
        '404':
          description: 'A Framework values collection not found'
...
  '/api/Framework values/{id}/{lang}':
    get:
      tags:
        - Framework
      summary: 'Display the specified framework value with specific language resource.'
      operationId: 'App\Http\Controllers\Api\Framework\ValueController::getOneWithLang'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: id
          in: path
          description: 'Numeric ID of the framework value to get'
          required: true
          schema:
            $ref: '#/components/schemas/resource_id'
        -
          name: lang
          in: path
          description: 'String for select a specific language'
          required: true
      responses:
        '200':
          description: 'One framework value'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/value_with_lang_resource'
        '404':
          description: 'A framework value not found'
...
  '/api/menus/{lang}':
    get:
      tags:
        - Menu
      summary: 'Display a collection of menus with specific language ressources.'
      operationId: 'App\Http\Controllers\Api\MenuController::getAllWithLang'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: lang
          in: path
          description: 'String for select a specific language'
          required: true
      responses:
        '200':
          description: 'A menu collection'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/menu_with_lang_collection_resource'
        '404':
          description: 'A menu collection not found'
...
  '/api/menus/{id}/{lang}':
    get:
      tags:
        - Menu
      summary: 'Display the specified menu with specific language resource.'
      operationId: 'App\Http\Controllers\Api\MenuController::getOneWithLang'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: id
          in: path
          description: 'Numeric ID of the menu to get'
          required: true
          schema:
            $ref: '#/components/schemas/resource_id'
        -
          name: lang
          in: path
          description: 'String for select a specific language'
          required: true
      responses:
        '200':
          description: 'One menu'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/menu_with_lang_resource'
        '404':
          description: 'A menu not found'
...
  '/api/pages/{lang}':
    get:
      tags:
        - Page
      summary: 'Display a collection of pages with specific language ressources.'
      operationId: 'App\Http\Controllers\Api\PageController::getAllWithLang'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: lang
          in: path
          description: 'String for select a specific language'
          required: true
      responses:
        '200':
          description: 'A pages collection'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/page_with_lang_collection_resource'
        '404':
          description: 'A pages collection not found'
...
  '/api/pages/{id}/{lang}':
    get:
      tags:
        - Page
      summary: 'Display the specified page with specific language resource.'
      operationId: 'App\Http\Controllers\Api\PageController::getOneWithLang'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: id
          in: path
          description: 'Numeric ID of the page to get'
          required: true
          schema:
            $ref: '#/components/schemas/resource_id'
        -
          name: lang
          in: path
          description: 'String for select a specific language'
          required: true
      responses:
        '200':
          description: 'One page'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/page_with_lang_resource'
        '404':
          description: 'A page not found'
...
  '/api/regions/{lang}':
    get:
      tags:
        - Region
      summary: 'Display a collection of regions with specific language ressources.'
      operationId: 'App\Http\Controllers\Api\RegionController::getAllWithLang'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: lang
          in: path
          description: 'String for select a specific language'
          required: true
      responses:
        '200':
          description: 'A regions collection'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/region_with_lang_collection_resource'
        '404':
          description: 'A regions collection not found'
...
  '/api/regions/{id}/{lang}':
    get:
      tags:
        - Region
      summary: 'Display the specified region with specific language resource.'
      operationId: 'App\Http\Controllers\Api\RegionController::getOneWithLang'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: id
          in: path
          description: 'Numeric ID of the region to get'
          required: true
          schema:
            $ref: '#/components/schemas/resource_id'
        -
          name: lang
          in: path
          description: 'String for select a specific language'
          required: true
      responses:
        '200':
          description: 'One region'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/region_with_lang_resource'
        '404':
          description: 'A region not found'
...
  '/api/teams/{lang}':
    get:
      tags:
        - Team
      summary: 'Display a collection of teams with specific language ressources.'
      operationId: 'App\Http\Controllers\Api\TeamController::getAllWithLang'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: lang
          in: path
          description: 'String for select a specific language'
          required: true
      responses:
        '200':
          description: 'A teams collection'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/team_with_lang_collection_resource'
        '404':
          description: 'A teams collection not found'
...
  '/api/teams/{id}/{lang}':
    get:
      tags:
        - Team
      summary: 'Display the specified team with specific language resource.'
      operationId: 'App\Http\Controllers\Api\TeamController::getOneWithLang'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: id
          in: path
          description: 'Numeric ID of the team to get'
          required: true
          schema:
            $ref: '#/components/schemas/resource_id'
        -
          name: lang
          in: path
          description: 'String for select a specific language'
          required: true
      responses:
        '200':
          description: 'One team'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/team_with_lang_resource'
        '404':
          description: 'A team not found'
...
  '/api/use-cases/{lang}':
    get:
      tags:
        - 'Use Case'
      summary: 'Display a collection of use cases with specific language ressources.'
      operationId: 'App\Http\Controllers\Api\UseCaseController::getAllWithLang'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: lang
          in: path
          description: 'String for select a specific language'
          required: true
      responses:
        '200':
          description: 'A use cases collection'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/use_case_with_lang_collection_resource'
        '404':
          description: 'A use cases collection not found'
...
  '/api/use-cases/{id}/{lang}':
    get:
      tags:
        - 'Use Case'
      summary: 'Display the specified use case with specific language resource.'
      operationId: 'App\Http\Controllers\Api\UseCaseController::getOneWithLang'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: id
          in: path
          description: 'Numeric ID of the office to get'
          required: true
          schema:
            $ref: '#/components/schemas/resource_id'
        -
          name: lang
          in: path
          description: 'String for select a specific language'
          required: true
      responses:
        '200':
          description: 'One use case'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/use_case_with_lang_resource'
        '404':
          description: 'A use case not found'
...
components:
  schemas:
    resource_id:
      description: 'The unique identifier of a resource'
      minimum: 1
      example: '1'
...
    certification_with_lang_collection_resource:
      required:
        - link
        - data
      properties:
        data:
          description: 'Transform the resource into an array.'
          type: array
          items:
            $ref: '#/components/schemas/certification_with_lang_resource'
        links:
          description: 'Transform the resource into an array.'
          required:
            - self
          properties:
            self:
              description: 'Transform the resource into an array.'
              type: string
              format: uri
              example: /api/framework/certifications/fr
          type: object
      type: object
...
    expertise_with_lang_collection_resource:
      required:
        - link
        - data
      properties:
        data:
          description: 'Transform the resource into an array.'
          type: array
          items:
            $ref: '#/components/schemas/expertise_with_lang_resource'
        links:
          description: 'Transform the resource into an array.'
          required:
            - self
          properties:
            self:
              description: 'Transform the resource into an array.'
              type: string
              format: uri
              example: /api/expertises/fr
          type: object
      type: object
    expertise_with_lang_resource:
      required:
        - type
        - id
        - attributes
        - links
      properties:
        type:
          description: 'Transform the resource into an array.'
          type: string
          example: expertises
        id:
          $ref: '#/components/schemas/resource_id'
        attributes:
          description: 'Transform the resource into an array.'
          required:
            - name
            - slug
            - subtitle
            - description
          properties:
            name:
              description: 'The name in specific language'
              type: string
              example: Delivrer
            slug:
              description: 'The slug in specific language'
              type: string
              example: delivrer
            subtitle:
              description: 'The subtitle in specific language'
              type: string
              example: 'Assurer la transition vers l''agile...'
            description:
              description: 'The description in specific language'
              type: string
              example: 'Installer la cross disciplinary d''une Digital Factory'
            color:
              description: 'The color in specific language'
              type: string
              example: '#123456'
          type: object
        links:
          description: 'Transform the resource into an array.'
          required:
            - self
          properties:
            self:
              description: 'Transform the resource into an array.'
              type: string
              format: uri
              example: /api/expertises/1
          type: object
      type: object
    formule_collection_resource:
      required:
        - link
        - data
      properties:
        data:
          description: 'Transform the resource into an array.'
          type: array
          items:
            $ref: '#/components/schemas/formule_resource'
        links:
          description: 'Transform the resource into an array.'
          required:
            - self
          properties:
            self:
              description: 'Transform the resource into an array.'
              type: string
              format: uri
              example: /api/formules
          type: object
      type: object
    formule_resource:
      required:
        - type
        - id
        - attributes
        - links
      properties:
        type:
          description: 'Transform the resource into an array.'
          type: string
          example: formules
        id:
          $ref: '#/components/schemas/resource_id'
        attributes:
          description: 'Transform the resource into an array.'
          required:
            - fr
            - en
          properties:
            fr:
              description: 'Transform the resource into an array.'
              required:
                - title
                - subtitle
                - price
                - options
              properties:
                title:
                  description: 'The title in specific language'
                  type: string
                  example: Delivrer
                subtitle:
                  description: 'The subtitle in specific language'
                  type: string
                  example: 'Assurer la transition vers l''agile...'
                price:
                  description: 'The price in specific language'
                  type: string
                  example: '450'
                options:
                  description: 'Transform the resource into an array.'
                  required: [option_1, option_12]
                  properties: { option_1: { description: 'The option_1 in specific language', type: boolean, example: 'true' }, option_12: { description: 'The option_12 in specific language', type: boolean, example: 'false' } }
                  type: object
              type: object
            en:
              description: 'Transform the resource into an array.'
              required:
                - title
                - subtitle
                - price
                - options
              properties:
                title:
                  description: 'The title in specific language'
                  type: string
                  example: Delivrer
                subtitle:
                  description: 'The subtitle in specific language'
                  type: string
                  example: 'Assurer la transition vers l''agile...'
                price:
                  description: 'The price in specific language'
                  type: string
                  example: '450'
                options:
                  description: 'Transform the resource into an array.'
                  required: [option_1, option_12]
                  properties: { option_1: { description: 'The option_1 in specific language', type: boolean, example: 'true' }, option_12: { description: 'The option_12 in specific language', type: boolean, example: 'false' } }
                  type: object
              type: object
          type: object
        links:
          description: 'Transform the resource into an array.'
          required:
            - self
          properties:
            self:
              description: 'Transform the resource into an array.'
              type: string
              format: uri
              example: /api/formules/1
          type: object
      type: object
    formule_with_lang_collection_resource:
      required:
        - link
        - data
      properties:
        data:
          description: 'Transform the resource into an array.'
          type: array
          items:
            $ref: '#/components/schemas/formule_with_lang_resource'
        links:
          description: 'Transform the resource into an array.'
          required:
            - self
          properties:
            self:
              description: 'Transform the resource into an array.'
              type: string
              format: uri
              example: /api/formules/fr
          type: object
      type: object
...
  securitySchemes:
    bearer:
      type: http
      name: bearerAuth
      bearerFormat: JWT
      scheme: bearer